added Creating tasks. fixed bugs on inviting users. added back getTasks  that returns all the tasks from every group you are in. also added getting users who are in your group.

+++ src/main/java/com/example/student/groups/service/GroupService.java
@@ -2,6 +2,7 @@ package com.example.student.groups.service;
 
 import com.example.student.groups.exceptions.AlreadyExistsException;
 import com.example.student.groups.exceptions.NotFoundException;
+import com.example.student.groups.exceptions.UnauthorizedException;
 import com.example.student.groups.model.Invites;
 import com.example.student.groups.model.TribeGroup;
 import com.example.student.student.TribeUser;
@@ -10,11 +11,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class GroupService {
 
     private final GroupRepository groupRepository;
+
     private final UserRepository userRepository;
 
     private final InviteRepository inviteRepository;
@@ -32,12 +35,31 @@ public class GroupService {
         }
 
         TribeUser user = findUserByFirebaseId(group.getAdminId());
-        group.addMember(user);
-
         TribeGroup savedTribeGroup = groupRepository.save(group);
 
         user.addGroup(savedTribeGroup.getId());
         userRepository.save(user);
+    }
+
+
+    public List<TribeGroup> getGroups() {
+        return groupRepository.findAll();
+    }
+
+    public void removeUserFromGroup(String adminFirebaseId, String userEmail) throws Exception {
+        TribeGroup group = getGroupByAdminId(adminFirebaseId);
+        TribeUser userToRemove = getUserByEmail(userEmail);
+
+        if (userToRemove.getGroups().contains(group.getId())) {
+            userToRemove.removeGroup(group.getId());
+        } else {
+            throw new NotFoundException("user is not in your group");
+        }
+        userRepository.save(userToRemove);
+    }
+
+    public List<Invites> getInvites(String firebaseId) {
+        return inviteRepository.findByUserFirebaseId(firebaseId);
     }
 
     public void inviteUserToGroup(String adminFirebaseId, String userEmail) throws Exception {
@@ -48,55 +70,38 @@ public class GroupService {
             throw new AlreadyExistsException("User already belongs to this group");
         }
 
-        Invites newInvite = new Invites(group.getTribeName(), group.getTribeDescription(), userToAdd.getFirebaseId());
+        Invites newInvite = new Invites(group.getTribeName(), group.getTribeDescription(), userToAdd.getFirebaseId(), group.getId());
         inviteRepository.save(newInvite);
     }
 
-    public List<Invites> getInvites(String firebaseId) {
-        return inviteRepository.findByUserFirebaseId(firebaseId);
-    }
-
-//    public List<TribeUser> getUsersInGroup(String firebaseId) throws Exception {
-//        TribeUser user = userRepository.findUserByFirebaseId(firebaseId)
-//                .orElseThrow(() -> new NotFoundException("User not found"));
-//        if (user.getGroupId() == null) {
-//            throw new NotFoundException("User is not in any group");
-//        }
-//        return userRepository.findByGroupId(user.getGroupId());
-//    }
-    public List<TribeGroup> getGroups() {
-        return groupRepository.findAll();
-    }
-
-    public void removeUserFromGroup(String adminFirebaseId, String userEmail) throws Exception {
-        TribeGroup group = getGroupByAdminId(adminFirebaseId);
-        TribeUser userToRemove = getUserByEmail(userEmail);
-
-        if (userToRemove.getGroups().contains(group.getId())){
-            userToRemove.removeGroup(group.getId());
-        }else {
-            throw new NotFoundException("user is not in your group");
-        }
-        userRepository.save(userToRemove);
-    }
-
-    public void invite(Long id, Integer accept) throws NotFoundException {
-        Invites invite = inviteRepository.findInviteById(id).orElseThrow(() -> new NotFoundException(""));
-        if(accept == 1){
-            addUserToGroup(id,invite.getUserFirebaseId());
+    public void invite(Long id, Integer accept, String firebaseId) throws NotFoundException {
+        Invites invite = inviteRepository.findInviteById(id).orElseThrow(() -> new NotFoundException("invite not found"));
+        if (!Objects.equals(invite.getUserFirebaseId(), firebaseId))
+            throw new NotFoundException("user with this invite not found");
+        if (accept == 1) {
+            addUserToGroup(invite.getGroupId(), invite.getUserFirebaseId());
             inviteRepository.delete(invite);
-        }else if (accept == -1){
+        } else if (accept == -1) {
             inviteRepository.delete(invite);
         }
     }
 
-    private void addUserToGroup(Long id,String firebaseId) throws NotFoundException {
+    private void addUserToGroup(Long id, String firebaseId) throws NotFoundException {
         TribeGroup group = groupRepository.getGroupById(id)
                 .orElseThrow(() -> new NotFoundException(" ID does not match to group ID"));
         TribeUser userToAdd = findUserByFirebaseId(firebaseId);
         userToAdd.addGroup(group.getId());
         userRepository.save(userToAdd);
     }
+
+
+    public List<TribeUser> getUsersInGroup(String firebaseId, Long id) throws Exception {
+        TribeUser user = findUserByFirebaseId(firebaseId);
+        if (user.getGroups().contains(id)){
+            return userRepository.findByGroupsContaining(id);
+        }else throw new UnauthorizedException("you dont have access to this");
+    }
+
     private TribeGroup getGroupByAdminId(String adminFirebaseId) throws NotFoundException {
         return groupRepository.getGroupByAdminId(adminFirebaseId)
                 .orElseThrow(() -> new NotFoundException("Firebase ID does not match admin ID"));
